requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.20
chromadb==0.4.22
openai==1.12.0
streamlit==1.31.0
python-dotenv==1.0.0
pypdf==0.0.0
tiktoken==0.5.2
crewai==0.1.26
crewai-tools==0.0.8
